You can now set the service URL if using yourls shortener.
I've modified the option.html page and added a new <p> for the yourls
options.  You can now set the service URL and your Signature token.

This <p> gets shown and hidden depending on the URL Shortener selected.

I've also added an extra param to the shorten function in
lib/shortener_lib.js this contains the serviceUrl for yourls.