Fixing desktop notifications not showing correctly.
This one was hard to debug, and it feels like a bug in Chrome to me.
Apparently RegExp.leftContext and RegExp.rightContext weren't returning
the expected result inside desktop notifications. With only one notification
everything worked fine, however with more than one being shown things
started to get mixed and incorrect or empty content would be shown inside
the notification. The solution was to avoid using this properties.